// src/lib/dropdown-input/dropdown-input-item-events.interface.ts
import { DynamicComponentData } from '../common/dynamic-component-data.interface';

export interface DropdownInputItemsEvent {
    event: Event;
    index: number;
}

export interface DropdownInputItemsMouseEvent extends DropdownInputItemsEvent {
        event: MouseEvent;
}

export interface DropdownInputItemChosenEvent {
    component: any;
    index: number;
    matchText: string;
}